add a Plots tab in explore page (only enabled by qc flag)
